// dashboard/src/utils/hasura-api/generated/schemas/dropLogicalModelSelectPermissionStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DropLogicalModelSelectPermissionArgs } from './dropLogicalModelSelectPermissionArgs';
import type { DropLogicalModelSelectPermissionStepType } from './dropLogicalModelSelectPermissionStepType';

export interface DropLogicalModelSelectPermissionStep {
  type: DropLogicalModelSelectPermissionStepType;
  args: DropLogicalModelSelectPermissionArgs;
}