// dashboard/src/utils/hasura-api/generated/schemas/nativeQueryMetadataBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateLogicalModelSelectPermissionStep } from './createLogicalModelSelectPermissionStep';
import type { DropLogicalModelSelectPermissionStep } from './dropLogicalModelSelectPermissionStep';
import type { NativeQueryMetadataBulkOperationType } from './nativeQueryMetadataBulkOperationType';
import type { TrackLogicalModelStep } from './trackLogicalModelStep';
import type { TrackNativeQueryStep } from './trackNativeQueryStep';
import type { UntrackLogicalModelStep } from './untrackLogicalModelStep';
import type { UntrackNativeQueryStep } from './untrackNativeQueryStep';

export interface NativeQueryMetadataBulkOperation {
  type: NativeQueryMetadataBulkOperationType;
  resource_version: number;
  args: (
    | TrackLogicalModelStep
    | UntrackLogicalModelStep
    | CreateLogicalModelSelectPermissionStep
    | DropLogicalModelSelectPermissionStep
    | TrackNativeQueryStep
    | UntrackNativeQueryStep
  )[];
}